# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/profiles": {
      "patch": {
        "operationId": "ProfilesController_updateProfile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Profile updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "success": true,
                    "message": "Profile updated successfully"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Validation error. Invalid field values",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": [
                      "name must be longer than or equal to 3 characters",
                      "phoneNumber must match phone format",
                      "gender must be one of: male, female",
                      "language must be one of: english, spanish",
                      "fitGoals must contain at least 1 elements"
                    ],
                    "error": "Bad Request",
                    "statusCode": 400
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Account suspended or deactivated"
          },
          "404": {
            "description": "Profile not found"
          }
        },
        "summary": "Update own profile",
        "tags": [
          "Profiles"
        ]
      }
    },
    "/profiles/complete-setup": {
      "post": {
        "operationId": "ProfilesController_completeProfileSetup",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Profile created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "success": true,
                    "message": "Profile created successfully",
                    "data": {
                      "name": "Jane Doe",
                      "birthDate": "1990-01-01",
                      "phoneNumber": "5512345678",
                      "gender": "female",
                      "fitGoals": [
                        "lose_weight"
                      ],
                      "language": "spanish",
                      "email": "user@example.com",
                      "createdAt": "2024-01-01T12:00:00Z",
                      "updatedAt": "2024-01-01T12:00:00Z",
                      "age": 34,
                      "currentStreak": 1,
                      "maxStreak": 1,
                      "lastWeight": 75.5,
                      "notificationPreferences": {
                        "routines": true,
                        "exercises": true,
                        "system": true
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Validation error or Profile already exists",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": [
                      "name must be longer than or equal to 3 characters",
                      "phoneNumber must match phone format",
                      "gender must be one of: male, female",
                      "language must be one of: english, spanish",
                      "fitGoals must contain at least 1 elements",
                      "Profile already exists for this account"
                    ],
                    "error": "Bad Request",
                    "statusCode": 400
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "summary": "Complete profile setup",
        "tags": [
          "Profiles"
        ]
      }
    },
    "/profiles/by-account/{accountId}": {
      "get": {
        "operationId": "ProfilesController_getProfileByAccountId",
        "parameters": [
          {
            "name": "accountId",
            "required": true,
            "in": "path",
            "schema": {
              "example": "uuid-v4",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Profile fetched successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "success": true,
                    "message": "Profile fetched successfully",
                    "data": {
                      "name": "Jane Doe",
                      "birthDate": "1990-01-01",
                      "phoneNumber": "5512345678",
                      "gender": "female",
                      "fitGoals": [
                        "lose_weight"
                      ],
                      "language": "english",
                      "email": "user@example.com",
                      "createdAt": "2024-01-01T12:00:00Z",
                      "updatedAt": "2024-01-02T12:00:00Z",
                      "age": 34,
                      "currentStreak": 5,
                      "maxStreak": 10,
                      "lastWeight": 75.5,
                      "notificationPreferences": {
                        "routines": true,
                        "exercises": true,
                        "system": true
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Profile not found"
          }
        },
        "summary": "Get profile by account ID",
        "tags": [
          "Profiles"
        ]
      }
    },
    "/profiles/me": {
      "get": {
        "operationId": "ProfilesController_findMyProfile",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Profile fetched successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "success": true,
                    "message": "Profile fetched successfully",
                    "data": {
                      "name": "Jane Doe",
                      "birthDate": "1990-01-01",
                      "phoneNumber": "5512345678",
                      "gender": "female",
                      "fitGoals": [
                        "lose_weight",
                        "build_muscle"
                      ],
                      "language": "english",
                      "email": "user@example.com",
                      "createdAt": "2024-01-01T12:00:00Z",
                      "updatedAt": "2024-01-02T12:00:00Z",
                      "age": 34,
                      "currentStreak": 5,
                      "maxStreak": 10,
                      "lastWeight": 75.5,
                      "notificationPreferences": {
                        "routines": true,
                        "exercises": true,
                        "system": true
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Profile not found"
          }
        },
        "summary": "Get current user profile",
        "tags": [
          "Profiles"
        ]
      }
    },
...
    "/Notifications": {
      "get": {
        "operationId": "NotificationsController_getUserNotifications",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Notifications retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "success": true,
                    "message": "Notifications retrieved successfully",
                    "data": {
                      "items": [
                        {
                          "id": "123e4567-e89b-12d3-a456-426614174000",
                          "title": "Rutina completada",
                          "message": "Has completado tu rutina de fuerza.",
                          "type": "success",
                          "isRead": false,
                          "createdAt": "2024-02-03T12:00:00Z",
                          "userId": "123e4567-e89b-12d3-a456-426614174001"
                        }
                      ],
                      "total": 1
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get user notifications",
        "tags": [
          "Notifications"
        ]
      }
    },
    "/Notifications/unread-count": {
      "get": {
        "operationId": "NotificationsController_getUnreadCount",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Count retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "success": true,
                    "message": "Count retrieved successfully",
                    "data": {
                      "count": 5
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get unread notifications count",
        "tags": [
          "Notifications"
        ]
      }
    },
    "/Notifications/{id}/read": {
      "patch": {
        "operationId": "NotificationsController_markAsRead",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Marked as read successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "success": true,
                    "message": "Notification marked as read",
                    "data": {
                      "success": true
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Notification not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Mark notification as read",
        "tags": [
          "Notifications"
        ]
      }
    },
    "/Notifications/read-all": {
      "patch": {
        "operationId": "NotificationsController_markAllAsRead",
        "parameters": [],
        "responses": {
          "200": {
            "description": "All marked as read successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "success": true,
                    "message": "All notifications marked as read",
                    "data": {
                      "success": true
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Mark all notifications as read",
        "tags": [
          "Notifications"
        ]
      }
    },
...
  },
  "info": {
    "title": "AthletIA API",
    "description": "API documentation for AthletIA",
    "version": "0.0.1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "ProfileRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Full name",
            "example": "Jane Doe"
          },
          "birthDate": {
            "format": "date-time",
            "type": "string",
            "description": "Birth date in ISO format (user must be 18+)",
            "example": "1990-01-01"
          },
          "phoneNumber": {
            "type": "string",
            "description": "10-digit phone number",
            "example": "5512345678"
          },
          "language": {
            "type": "string",
            "description": "Language preference - Valid values: english, spanish",
            "enum": [
              "english",
              "spanish"
            ],
            "example": "spanish"
          },
          "gender": {
            "type": "string",
            "description": "Gender - Valid values: male, female",
            "enum": [
              "male",
              "female"
            ],
            "example": "male"
          },
          "fitGoals": {
            "type": "array",
            "description": "Fitness goals - Valid values: weight_loss, muscle_gain, weight_maintenance, endurance, flexibility, general_fitness, rehabilitation, improved_posture, balance_and_coordination, cardiovascular_health, strength_training, athletic_performance, lifestyle_enhancement",
            "example": [
              "weight_loss",
              "muscle_gain"
            ],
            "items": {
              "type": "string",
              "enum": [
                "weight_loss",
                "muscle_gain",
                "weight_maintenance",
                "endurance",
                "flexibility",
                "general_fitness",
                "rehabilitation",
                "improved_posture",
                "balance_and_coordination",
                "cardiovascular_health",
                "strength_training",
                "athletic_performance",
                "lifestyle_enhancement"
              ]
            }
          }
        },
        "required": [
          "name",
          "birthDate",
          "phoneNumber",
          "language",
          "gender",
          "fitGoals"
        ]
      },
...
      "NotificationPreferencesDto": {
        "type": "object",
        "properties": {
          "routines": {
            "type": "boolean",
            "description": "Receive routine notifications",
            "example": true
          },
          "exercises": {
            "type": "boolean",
            "description": "Receive exercise notifications",
            "example": true
          },
          "system": {
            "type": "boolean",
            "description": "Receive system notifications",
            "example": true
          }
        }
      },
...
      "ProfileUpdate": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Full name",
            "example": "Jane Doe"
          },
          "birthDate": {
            "format": "date-time",
            "type": "string",
            "description": "Birth date in ISO format (user must be 18+)",
            "example": "1990-01-01"
          },
          "phoneNumber": {
            "type": "string",
            "description": "10-digit phone number",
            "example": "5512345678"
          },
          "language": {
            "type": "string",
            "description": "Language preference - Valid values: english, spanish",
            "enum": [
              "english",
              "spanish"
            ],
            "example": "spanish"
          },
          "notificationPreferences": {
            "description": "Notification preferences",
            "allOf": [
              {
                "$ref": "#/components/schemas/NotificationPreferencesDto"
              }
            ]
          }
        }
      },
...
    }
  }
}